Use long options to docker command
When demonstrating example commands to run (which will likely be copy/pasted into a script or a shell directly), it's much more instructive to use the long options. Even if a user isn't intimately familiar with docker, using the long options give them a chance to see what each one is doing rather than being a more impenetrable series of cryptic symbols.